Rework vimrc with help of vint
Change all double quoted strings to single quoted. Similar to shell, it
prevents expansion where you don't want it.
Remove nocompatible, since it isn't needed, and when run more than once,
it can mess up settings.
Move encoding to the first line, since it needs to be before
scriptencoding, which needs to be first.
Expand all set properties to their long form, for clarity.
Prefix all variables with their respective scope.
Wrap autocommands in an augroup, which prevents duplication on reload,
and an easy way to wipe out them all.
Rewrite HighlightRepeats function to use a simple regex, since the
old one was not working constantly. The new one only works on adjacent
lines, but for data instead of text, which is what it was intended for,
that's not an issue.
Rewrite TrimWhitespace function, since it used the %s command, which
isn't portable and will behave differently based on current settings.
The new one prompts for each line anyway, so it's safer.